requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pycodestyle==2.5.0
python-decouple==3.1
